Fix collisions
